+++ Client/Client.Qt/Widgets/IPopupMessage/IPopupMessage.hpp
@@ -2,18 +2,18 @@
 
 #include <QWidget>
 
-class I_PopupMessage : public QWidget
+class IPopupMessage : public QWidget
 {
     Q_OBJECT
     Q_PROPERTY(float popupOpacity READ getPopupOpacity WRITE setPopupOpacity)
 public:
 
-    I_PopupMessage(QWidget* parent)
+    IPopupMessage(QWidget* parent)
         : QWidget(parent) 
     {
     }
     
-    virtual ~I_PopupMessage(){};
+    virtual ~IPopupMessage(){};
     
     virtual void setPopupText(const QString& text)  = 0;
     virtual void  popupShow()                       = 0;
